Set max_input_tokens to default to None and move max_output_tokens tests to test_llm.py
